+++ src/Paket.Core/Common/Utils.fs
@@ -155,7 +155,23 @@ let getDirectoryInfoForLocalNuGetFeed pathInfo alternativeProjectRoot root =
         match alternativeProjectRoot with
         | Some root -> DirectoryInfo(Path.Combine(root, s))
         | None -> DirectoryInfo(Path.Combine(root, s))
-        
+
+
+// show the path that was too long
+let FileInfo(str) =
+    try
+        FileInfo(str)
+    with
+      :? PathTooLongException as exn -> raise (PathTooLongException("Path too long: " + str, exn))
+
+
+// show the path that was too long
+let DirectoryInfo(str) =
+    try
+        DirectoryInfo(str)
+    with
+      :? PathTooLongException as exn -> raise (PathTooLongException("Path too long: " + str, exn))
+
 /// Creates a directory if it does not exist.
 let createDir path = 
     try
@@ -786,20 +802,52 @@ let inline windowsPath (path:string) = path.Replace(Path.DirectorySeparatorChar,
 /// Gets all files with the given pattern
 let inline FindAllFiles(folder, pattern) = DirectoryInfo(folder).GetFiles(pattern, SearchOption.AllDirectories)
 
-let getTargetFolder root groupName (packageName:PackageName) version includeVersionInPath = 
-    let packageFolder = string packageName + if includeVersionInPath then "." + string version else ""
-    if groupName = Constants.MainDependencyGroup then
-        Path.Combine(root, Constants.PackagesFolderName, packageFolder)
-    else
-        Path.Combine(root, Constants.PackagesFolderName, groupName.CompareString, packageFolder)
+type ResolvedPackagesFolder =
+    /// No "packages" folder for the current package
+    | NoPackagesFolder
+    /// the /packages/group/ExtractedPackage.X.Y.Z folder
+    | ResolvedFolder of string
+    member x.Path =
+        match x with
+        | NoPackagesFolder -> None
+        | ResolvedFolder f -> Some f
 
+type PackagesFolderGroupConfig =
+    | NoPackagesFolder
+    | GivenPackagesFolder of string
+    | DefaultPackagesFolder
+    member x.ResolveGroupDir root groupName =
+        match x with
+        | NoPackagesFolder -> None
+        | GivenPackagesFolder p ->
+            // relative to root
+            Some p
+        | DefaultPackagesFolder ->
+            let groupDir =
+                if groupName = Constants.MainDependencyGroup then
+                    Path.Combine(root, Constants.DefaultPackagesFolderName)
+                else
+                    Path.Combine(root, Constants.DefaultPackagesFolderName, groupName.CompareString)
+            Some groupDir
+    member x.Resolve root groupName (packageName:PackageName) version includeVersionInPath  =
+        match x with
+        | NoPackagesFolder -> ResolvedPackagesFolder.NoPackagesFolder
+        | GivenPackagesFolder p ->
+            // relative to root
+            ResolvedPackagesFolder.ResolvedFolder p
+        | DefaultPackagesFolder ->
+            let groupDir = x.ResolveGroupDir root groupName |> Option.get
+            let packageFolder = string packageName + if includeVersionInPath then "." + string version else ""
+            let parent = Path.Combine(groupDir, packageFolder)
+            ResolvedPackagesFolder.ResolvedFolder parent
+    static member Default = DefaultPackagesFolder
 let RunInLockedAccessMode(rootFolder,action) =
-    let packagesFolder = Path.Combine(rootFolder,Constants.PackagesFolderName)
-    if Directory.Exists packagesFolder |> not then
-        Directory.CreateDirectory packagesFolder |> ignore
+    let paketFilesFolder = Path.Combine(rootFolder,Constants.PaketFilesFolderName)
+    if Directory.Exists paketFilesFolder |> not then
+        Directory.CreateDirectory paketFilesFolder |> ignore
 
     let p = System.Diagnostics.Process.GetCurrentProcess()
-    let fileName = Path.Combine(packagesFolder,Constants.AccessLockFileName)
+    let fileName = Path.Combine(paketFilesFolder,Constants.AccessLockFileName)
 
     // Checks the packagesFolder for a paket.locked file or waits until it get access to it.
     let rec acquireLock (startTime:DateTime) (timeOut:TimeSpan) trials =
@@ -1022,6 +1070,14 @@ let removeComment (text:string) =
         | -1, p | p , -1 -> stripComment p
         | p1, p2 -> stripComment (min p1 p2) 
     remove 0
+
+let getSha512Stream (stream:Stream) =
+    use hasher = System.Security.Cryptography.SHA512.Create() :> System.Security.Cryptography.HashAlgorithm
+    Convert.ToBase64String(hasher.ComputeHash(stream))
+
+let getSha512File (filePath:string) =
+    use stream = File.OpenRead(filePath)
+    getSha512Stream stream
 
 // adapted from MiniRx
 // http://minirx.codeplex.com/