+++ src/Paket.Core/PackageAnalysis/Queries.fs
@@ -9,7 +9,7 @@ let getLockFileFromDependenciesFile dependenciesFileName =
     let lockFileName = DependenciesFile.FindLockfile dependenciesFileName
     LockFile.LoadFrom lockFileName.FullName
 
-let listPackages (packages: System.Collections.Generic.KeyValuePair<GroupName*PackageName, PackageResolver.ResolvedPackage> seq) =
+let listPackages (packages: System.Collections.Generic.KeyValuePair<GroupName*PackageName, PackageResolver.PackageInfo> seq) =
     packages
     |> Seq.map (fun kv ->
             let groupName,packageName = kv.Key
@@ -23,18 +23,22 @@ let getInstalledPackageModel (lockFile: LockFile) (QualifiedPackageName(groupNam
     match lockFile.Groups |> Map.tryFind groupName with
     | None -> failwithf "Group %O can't be found in paket.lock." groupName
     | Some group ->
-        match group.Resolution.TryFind(packageName) with
+        match group.TryFind(packageName) with
         | None -> failwithf "Package %O is not installed in group %O." packageName groupName
         | Some resolvedPackage ->
             let packageName = resolvedPackage.Name
-            let folder = 
-                getTargetFolder lockFile.RootPath groupName packageName resolvedPackage.Version (defaultArg resolvedPackage.Settings.IncludeVersionInPath false)
-                |> Path.GetFullPath
+            let storage = defaultArg resolvedPackage.Settings.StorageConfig PackagesFolderGroupConfig.Default
+            let includeVersion = defaultArg resolvedPackage.Settings.IncludeVersionInPath false
+            let resolvedConfig = storage.Resolve lockFile.RootPath groupName packageName resolvedPackage.Version includeVersion
+            let folder =
+                match resolvedConfig.Path with
+                | Some f -> f
+                | None -> NuGetCache.GetTargetUserFolder packageName resolvedPackage.Version
 
             InstallModel.CreateFromContent(
                 packageName,
                 resolvedPackage.Version,
-                Paket.Requirements.FrameworkRestriction.NoRestriction, 
+                Paket.Requirements.FrameworkRestriction.NoRestriction,
                 NuGet.GetContent(folder).Force())
 
 let getRuntimeGraph (lockFile: LockFile) (groupName:GroupName) =