+++ src/Paket.Core/PackageManagement/NugetConvert.fs
@@ -458,7 +458,7 @@ let convertR rootDirectory force credsMigrationMode = trial {
     return! createResult(rootDirectory, nugetEnv, credsMigrationMode)
 }
 
-let replaceNuGetWithPaket initAutoRestore installAfter fromBootstrapper result = 
+let replaceNuGetWithPaket initAutoRestore installAfter result = 
     let remove (fi : FileInfo) = 
         tracefn "Removing %s" fi.FullName
         fi.Delete()
@@ -497,7 +497,7 @@ let replaceNuGetWithPaket initAutoRestore installAfter fromBootstrapper result =
     
     if initAutoRestore && (autoVSPackageRestore || result.NuGetEnv.NuGetTargets.IsSome) then
         try
-            VSIntegration.TurnOnAutoRestore fromBootstrapper result.PaketEnv |> returnOrFail
+            VSIntegration.TurnOnAutoRestore result.PaketEnv |> returnOrFail
         with
         | exn -> 
             traceWarnfn "Could not enable auto restore%sMessage: %s" Environment.NewLine exn.Message